Factor out weak_method.
